cr: Fix clang compiler warning
GIT_ORIGIN_SPP_REV_ID: 5ad4edec5fc0bbdced3fb28cbdc8c20a1808f643

+++ score/os/qnx/channel_impl.cpp
@@ -77,7 +77,7 @@ score::cpp::expected<std::int32_t, score::os::Error> ChannelImpl::MsgReceivePuls
 }
 
 /* KW_SUPPRESS_START:MISRA.VAR.HIDDEN:Wrapper function is identifiable through namespace usage */
-score::cpp::expected_blank<score::os::Error> ChannelImpl::MsgReply(const std::int32_t rcvid,
+score::cpp::expected_blank<score::os::Error> ChannelImpl::MsgReply(const std::int64_t rcvid,
                                                           const std::int64_t status,
                                                           const void* const msg,
                                                           const std::size_t bytes) const noexcept
