+++ score/os/qnx/channel_impl.h
@@ -51,7 +51,7 @@ class ChannelImpl final : public Channel
     /* KW_SUPPRESS_END:MISRA.USE.EXPANSION:Using library-defined macro to ensure correct operation */
 
     /* KW_SUPPRESS_START:MISRA.VAR.HIDDEN:Wrapper function is identifiable through namespace usage */
-    score::cpp::expected_blank<score::os::Error> MsgReply(const std::int32_t rcvid,
+    score::cpp::expected_blank<score::os::Error> MsgReply(const std::int64_t rcvid,
                                                  const std::int64_t status,
                                                  const void* const msg,
                                                  const std::size_t bytes) const noexcept override;
